Add files via upload
The application's Main Page/Home Page is the Login Page which takes input as email and password from the user and directs them to the next page after clicking on the Login button. It also provides the user with the functionality of continuing as guest as not all users might want to register in the first place. To provide the user with notifications, an alert window pops us after the user has clicked the respective links(either logged in or decided to continue as guest).